+++ src/SiteLinkList.php
@@ -9,9 +9,11 @@ use InvalidArgumentException;
 use IteratorAggregate;
 use OutOfBoundsException;
 use Traversable;
+use Wikibase\DataModel\Entity\ItemId;
+use Wikibase\DataModel\Entity\ItemIdSet;
 
 /**
- * Immutable unordered collection of SiteLink objects.
+ * Unordered collection of SiteLink objects.
  * SiteLink objects can be accessed by site id.
  * Only one SiteLink per site id can exist in the collection.
  *
@@ -34,16 +36,40 @@ class SiteLinkList implements IteratorAggregate, Countable, Comparable {
 				throw new InvalidArgumentException( 'SiteLinkList only accepts SiteLink objects' );
 			}
 
-			$this->addSiteLink( $siteLink );
+			$this->addObject( $siteLink );
 		}
 	}
 
-	private function addSiteLink( SiteLink $link ) {
+	/**
+	 * @since 0.8
+	 *
+	 * @param SiteLink $link
+	 *
+	 * @throws InvalidArgumentException
+	 * @return self
+	 */
+	public function addObject( SiteLink $link ) {
 		if ( array_key_exists( $link->getSiteId(), $this->siteLinks ) ) {
 			throw new InvalidArgumentException( 'Duplicate site id: ' . $link->getSiteId() );
 		}
 
 		$this->siteLinks[$link->getSiteId()] = $link;
+
+		return $this;
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @param string $siteId
+	 * @param string $pageName
+	 * @param ItemIdSet|ItemId[] $badges
+	 *
+	 * @throws InvalidArgumentException
+	 * @return self
+	 */
+	public function add( $siteId, $pageName, $badges = array() ) {
+		return $this->addObject( new SiteLink( $siteId, $pageName, $badges ) );
 	}
 
 	/**
@@ -67,19 +93,33 @@ class SiteLinkList implements IteratorAggregate, Countable, Comparable {
 
 	/**
 	 * @param string $siteId
+	 *
+	 * @return SiteLink
 	 * @throws OutOfBoundsException
 	 * @throws InvalidArgumentException
 	 */
 	public function getBySiteId( $siteId ) {
-		if ( !is_string( $siteId ) ) {
-			throw new InvalidArgumentException( '$siteId should be a string' );
-		}
-
-		if ( !array_key_exists( $siteId, $this->siteLinks ) ) {
+		if ( !$this->hasLinkWithSiteId( $siteId ) ) {
 			throw new OutOfBoundsException( 'No SiteLink with site id: ' . $siteId  );
 		}
 
 		return $this->siteLinks[$siteId];
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @param string $siteId
+	 *
+	 * @return boolean
+	 * @throws InvalidArgumentException
+	 */
+	public function hasLinkWithSiteId( $siteId ) {
+		if ( !is_string( $siteId ) ) {
+			throw new InvalidArgumentException( '$siteId should be a string' );
+		}
+
+		return array_key_exists( $siteId, $this->siteLinks );
 	}
 
 	/**
@@ -97,6 +137,29 @@ class SiteLinkList implements IteratorAggregate, Countable, Comparable {
 		}
 
 		return $this->siteLinks == $target->siteLinks;
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @return boolean
+	 */
+	public function isEmpty() {
+		return empty( $this->siteLinks );
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @param string $siteId
+	 * @throws InvalidArgumentException
+	 */
+	public function removeLinkWithSiteId( $siteId ) {
+		if ( !is_string( $siteId ) ) {
+			throw new InvalidArgumentException( '$siteId should be a string' );
+		}
+
+		unset( $this->siteLinks[$siteId] );
 	}
 
 }