+++ src/Term/AliasGroupList.php
@@ -31,7 +31,7 @@ class AliasGroupList implements Countable, IteratorAggregate {
 	 * @param AliasGroup[] $aliasGroups
 	 * @throws InvalidArgumentException
 	 */
-	public function __construct( array $aliasGroups ) {
+	public function __construct( array $aliasGroups = array() ) {
 		foreach ( $aliasGroups as $aliasGroup ) {
 			if ( !( $aliasGroup instanceof AliasGroup ) ) {
 				throw new InvalidArgumentException( 'AliasGroupList can only contain AliasGroup instances' );
@@ -142,7 +142,29 @@ class AliasGroupList implements Countable, IteratorAggregate {
 	 */
 	public function hasAliasGroup( AliasGroup $group ) {
 		return array_key_exists( $group->getLanguageCode(), $this->groups )
-			&& $this->groups[$group->getLanguageCode()]->equals( $group );
+		&& $this->groups[$group->getLanguageCode()]->equals( $group );
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @param string $languageCode
+	 *
+	 * @return boolean
+	 */
+	public function hasGroupForLanguage( $languageCode ) {
+		$this->assertIsLanguageCode( $languageCode );
+		return array_key_exists( $languageCode, $this->groups );
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @param string $languageCode
+	 * @param string[] $aliases
+	 */
+	public function setAliasesForLanguage( $languageCode, array $aliases ) {
+		$this->setGroup( new AliasGroup( $languageCode, $aliases ) );
 	}
 
 }