+++ src/Term/TermList.php
@@ -29,7 +29,7 @@ class TermList implements Countable, IteratorAggregate, Comparable {
 	 * @param Term[] $terms
 	 * @throws InvalidArgumentException
 	 */
-	public function __construct( array $terms ) {
+	public function __construct( array $terms = array() ) {
 		foreach ( $terms as $term ) {
 			if ( !( $term instanceof Term ) ) {
 				throw new InvalidArgumentException( 'TermList can only contain instances of Term' );
@@ -99,7 +99,7 @@ class TermList implements Countable, IteratorAggregate, Comparable {
 		return array_key_exists( $languageCode, $this->terms );
 	}
 
-	protected function assertIsLanguageCode( $languageCode ) {
+	private function assertIsLanguageCode( $languageCode ) {
 		if ( !is_string( $languageCode ) ) {
 			throw new InvalidArgumentException( '$languageCode should be a string' );
 		}
@@ -107,6 +107,16 @@ class TermList implements Countable, IteratorAggregate, Comparable {
 
 	public function setTerm( Term $term ) {
 		$this->terms[$term->getLanguageCode()] = $term;
+	}
+
+	/**
+	 * @since 0.8
+	 *
+	 * @param string $languageCode
+	 * @param string $termText
+	 */
+	public function setTextForLanguage( $languageCode, $termText ) {
+		$this->setTerm( new Term( $languageCode, $termText ) );
 	}
 
 	/**
@@ -145,7 +155,7 @@ class TermList implements Countable, IteratorAggregate, Comparable {
 	 */
 	public function hasTerm( Term $term ) {
 		return array_key_exists( $term->getLanguageCode(), $this->terms )
-			&& $this->terms[$term->getLanguageCode()]->equals( $term );
+		&& $this->terms[$term->getLanguageCode()]->equals( $term );
 	}
 
 }